Move Integration details into separate dir
