+++ app/ui-react/packages/ui/src/Integration/Editor/IntegrationEditorForm.tsx
@@ -1,7 +1,7 @@
 import * as H from '@syndesis/history';
 import { Alert } from 'patternfly-react';
 import * as React from 'react';
-import { ButtonLink, Container, PageSection } from '../Layout';
+import { ButtonLink, Container, PageSection } from '../../Layout';
 
 export interface IIntegrationEditorFormProps {
   /**
