+++ app/ui-react/packages/ui/src/Integration/Editor/endpoint/DescribeDataShapeForm.tsx
