+++ app/ui-react/packages/ui/src/Integration/Editor/template/TemplateStepCard.tsx
