+++ app/ui-react/packages/ui/src/Integration/index.ts
@@ -1,18 +1,9 @@
-export * from './editor';
-export * from './CiCd';
-export * from './Import';
-export * from './Details';
 export * from './Activity';
+export * from './CiCd';
+export * from './Details';
+export * from './Editor';
+export * from './Import';
 export * from './Metrics';
-export * from './IntegrationEditorActionsListItem';
-export * from './IntegrationEditorChooseAction';
-export * from './IntegrationEditorConnectionsListItem';
-export * from './IntegrationEditorForm';
-export * from './IntegrationEditorLayout';
-export * from './IntegrationEditorNothingToConfigure';
-export * from './IntegrationEditorStepsList';
-export * from './IntegrationEditorStepsListItem';
-export * from './IntegrationEditorWizard';
 export * from './IntegrationFlowAddStep';
 export * from './IntegrationFlowStep';
 export * from './IntegrationFlowStepDetails';
